Add missing cellular configuration items.
In Mbed OS 5.11.5 and older, the cellular does not have configuration
items for cellular pins and device name.

In Mbed OS 5.12 and later, the cellular configuration is refactored
to have appropriate mbed_lib.json defines. And because of that, the
configuration items included in this PR SHOULD NOT be used with
Mbed OS 5.12.